% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{heatmap.simulations}
\alias{heatmap.simulations}
\title{Multiple simulations for a heatmap plot}
\usage{
heatmap.simulations(res, no.draws = 1000)
}
\arguments{
\item{res}{A bard result}

\item{no.draws}{Number of draws to take from posterior distribution}
}
\value{
A list of matrices
}
\description{
Multiple simulations for a heatmap plot
}
\examples{
set.seed(1)
k_N <- 10
p_N <- 0.1
k_A <- 15
p_A <- 0.3
pi_N <- 0.8
affected_dim <- 8
lower <- 0.3
upper <- 0.7
simparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim, lower, upper)
sim <- sim.data(T = 1000, N = 200, simparams)

bardparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim)
mu_seq <- c(seq(-0.7, -0.3, by=0.05), seq(0.3, 0.7, by=0.05))
res <- Rbard(sim$data, bardparams, mu_seq, alpha = 1e-4)

hs = heatmap.simulations(res, no.draws = 100)

}
