\name{iacounty.sdf}
\alias{iacounty.sdf}
\docType{data}
\title{
suffolk map 
}
\description{
county level map data
}

\usage{data(iacounty.sdf)}
\format{
	Format is a SpatialPolygonsDataFrame
}
\details{
	A spatial data frame object of 2000 Iowa counties
}
\examples{
data(iacounty.sdf)
}
\keyword{datasets}
