% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{marginal.prob}
\alias{marginal.prob}
\title{Marginal probability of being abnormal}
\usage{
marginal.prob(res, no.draws = 1000, plot = F)
}
\arguments{
\item{res}{A bard result}

\item{no.draws}{Number of draws to take from posterior distribution}

\item{plot}{If true produce plot}
}
\value{
Marginal probability of being abnormal at each point
}
\description{
Marginal probability of being abnormal
}
\examples{
set.seed(1)
k_N <- 10
p_N <- 0.1
k_A <- 15
p_A <- 0.3
pi_N <- 0.8
affected_dim <- 8
lower <- 0.3
upper <- 0.7
simparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim, lower, upper)
sim <- sim.data(T = 1000, N = 200, simparams)

bardparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim)
mu_seq <- c(seq(-0.7, -0.3, by=0.05), seq(0.3, 0.7, by=0.05))
res <- Rbard(sim$data, bardparams, mu_seq, alpha = 1e-4)

margprob = marginal.prob(res, no.draws = 100, plot = T)

}
