% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bard.R
\name{Rbard}
\alias{Rbard}
\title{Compute BARD recursions}
\usage{
Rbard(data, bardparams, mu_seq, alpha = 1e-04)
}
\arguments{
\item{data}{T by N matrix}

\item{bardparams}{Vector of length 6 containing  k_N, p_N, k_A, p_A, pi_N, affected_dim}

\item{mu_seq}{Prior for mu}

\item{alpha}{Parameter used in resampling step}
}
\value{
List containing
}
\description{
Compute BARD recursions
}
\examples{
set.seed(1)
k_N <- 10
p_N <- 0.1
k_A <- 15
p_A <- 0.3
pi_N <- 0.8
affected_dim <- 8
lower <- 0.3
upper <- 0.7
simparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim, lower, upper)
sim <- sim.data(T = 1000, N = 200, simparams)

bardparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim)
mu_seq <- c(seq(-0.7, -0.3, by=0.05), seq(0.3, 0.7, by=0.05))
res <- Rbard(sim$data, bardparams, mu_seq, alpha = 1e-4)

}
