% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim.data}
\alias{sim.data}
\title{Simulate data from model}
\usage{
sim.data(T = 1000, N = 200, simparams = NA)
}
\arguments{
\item{T}{Length of series}

\item{N}{Number of variates}

\item{simparams}{Vector of length 8 containing  k_N, p_N, k_A, p_A, pi_N, affected_dim, lower and upper}
}
\value{
List containing data (T by N matrix), states and changepoints
}
\description{
Simulate data from model
}
\examples{
set.seed(1)
k_N <- 10
p_N <- 0.1
k_A <- 15
p_A <- 0.3
pi_N <- 0.8
affected_dim <- 8
lower <- 0.3
upper <- 0.7
simparams <- c(k_N, p_N, k_A, p_A, pi_N, affected_dim, lower, upper)
sim <- sim.data(T = 1000, N = 200, simparams)

}
